requirements.txt
dvc
numpy
scipy
requests
discord
python-dotenv
#keyring